+++ changelog.md
@@ -34,6 +34,7 @@
       print(x) -- `x` is `string` here
   end
   ```
+* `CHG` infer type by `>`/`<`/`>=`/`<=`
 * `FIX` with clients that support LSP 3.17 (VSCode), workspace diagnostics are triggered every time when opening a file.
 * `FIX` [#1204](https://github.com/sumneko/lua-language-server/issues/1204)
 * `FIX` [#1208](https://github.com/sumneko/lua-language-server/issues/1208)